Some tidying up of tests.
